+++ src/main/java/edu/ie3/datamodel/io/source/TimeSeriesSource.java
@@ -5,22 +5,39 @@
 */
 package edu.ie3.datamodel.io.source;
 
-import edu.ie3.datamodel.models.timeseries.TimeSeriesContainer;
-import edu.ie3.datamodel.models.timeseries.mapping.TimeSeriesMapping;
-import java.util.Set;
+import edu.ie3.datamodel.models.timeseries.individual.IndividualTimeSeries;
+import edu.ie3.datamodel.models.value.Value;
+import edu.ie3.util.interval.ClosedInterval;
+import java.time.ZonedDateTime;
+import java.util.Optional;
 
-public interface TimeSeriesSource {
-  /**
-   * Receive a set of time series mapping entries from participant uuid to time series uuid.
-   *
-   * @return A set of time series mapping entries from participant uuid to time series uuid
-   */
-  Set<TimeSeriesMapping.Entry> getMapping();
+/**
+ * The interface definition of a source, that is able to provide one specific time series for one
+ * model
+ */
+public interface TimeSeriesSource<V extends Value> extends DataSource {
 
   /**
-   * Acquire all available time series
+   * Obtain the full time series
    *
-   * @return A container with all relevant time series
+   * @return the time series
    */
-  TimeSeriesContainer getTimeSeries();
+  IndividualTimeSeries<V> getTimeSeries();
+
+  /**
+   * Get the time series for the given time interval. If the interval is bigger than the time series
+   * itself, only the parts of the time series within the interval are handed back.
+   *
+   * @param timeInterval Desired time interval to cover
+   * @return The parts of of interest of the time series
+   */
+  IndividualTimeSeries<V> getTimeSeries(ClosedInterval<ZonedDateTime> timeInterval);
+
+  /**
+   * Get the time series value for a specific time
+   *
+   * @param time The queried time
+   * @return Option on a value for that time
+   */
+  Optional<V> getValue(ZonedDateTime time);
 }