+++ src/test/groovy/edu/ie3/datamodel/io/sink/InfluxDbSinkIT.groovy
@@ -34,7 +34,7 @@ import java.time.ZonedDateTime
 class InfluxDbSinkIT extends Specification {
 
 	@Shared
-	InfluxDBContainer influxDbContainer = new InfluxDBContainer("latest")
+	InfluxDBContainer influxDbContainer = new InfluxDBContainer("1.8.4")
 	.withAuthEnabled(false)
 	.withDatabase("test_out")
 	.withExposedPorts(8086) as InfluxDBContainer
@@ -108,11 +108,14 @@ class InfluxDbSinkIT extends Specification {
 		def chpResult1 = new ChpResult(ZonedDateTime.of(2020, 5, 3, 14, 18, 0, 0, ZoneId.of("UTC")),
 				UUID.randomUUID(),
 				Quantities.getQuantity(42.24, StandardUnits.ACTIVE_POWER_RESULT),
-				Quantities.getQuantity(-42.24, StandardUnits.REACTIVE_POWER_RESULT))
+				Quantities.getQuantity(-42.24, StandardUnits.REACTIVE_POWER_RESULT),
+				Quantities.getQuantity(1.01, StandardUnits.Q_DOT_RESULT)
+				)
 		def chpResult2 = new ChpResult(ZonedDateTime.of(2020, 5, 3, 14, 19, 0, 0, ZoneId.of("UTC")),
 				UUID.randomUUID(),
 				Quantities.getQuantity(24.42, StandardUnits.ACTIVE_POWER_RESULT),
-				Quantities.getQuantity(-24.42, StandardUnits.REACTIVE_POWER_RESULT))
+				Quantities.getQuantity(-24.42, StandardUnits.REACTIVE_POWER_RESULT),
+				Quantities.getQuantity(1.01, StandardUnits.Q_DOT_RESULT))
 		def entities = [
 			lineResult1,
 			lineResult2,